python-crawler/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
sqlalchemy
pydantic-settings
tenacity
pytest
httpx
trafilatura
spacy
scikit-learn
numpy
uvicorn[standard]
fastapi
python-multipart
